+++ sdk/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.0] - 2024-06-07
+
 ### Added
 
 - `Client` gets Release resources from GitHub `giantswarm/releases` repo
@@ -26,5 +28,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Release` type that represents a Giant Swarm workload cluster release.
 
-[Unreleased]: https://github.com/giantswarm/releases/compare/sdk/v0.1.0...HEAD
+[Unreleased]: https://github.com/giantswarm/releases/compare/sdk/v0.2.0...HEAD
+[0.2.0]: https://github.com/giantswarm/releases/releases/tag/sdk/v0.2.0
 [0.1.0]: https://github.com/giantswarm/releases/releases/tag/sdk/v0.1.0